lua/common_string.lua: Improve split_string()
Check if the split char is even in the input string and if it is
not, then just add the original string to the buffer and return
that.

Add a "keep" flag which if true means to add the split char to the
buffer as a separate item each time it is encountered.

Signed-off-by: James Carter <jwcarter@gmail.com>