lua: Improve buffer operations
Make the format data a part of the buffer that is created when
creating a new buffer.

Change the names of the buffer functions to start with "buffer_".

Split lines when adding to the buffer and wrap lines when writing
the buffer. This means that the padding will have already been
applied and will be taken into account when wrapping the line.
Whereas before the line length was checked before wrapping, so
a line could end up longer than desired once the wrapping was
applied.

Signed-off-by: James Carter <jwcarter@gmail.com>