lua/simpl_write.lua: Write rules in a reproducible order
A few functions were using "for key,data in pairs(data_table)" when
writing rules which causes the order of the rules to change.

Instead, first sort the keys of the table and then walk the sorted
list using "for i=1,#keys do" so the order the rules are written
will always be the same.

Signed-off-by: James Carter <jwcarter@gmail.com>