lua/simpl_write.lua: Fix problem with replacing compound args
SIMPL does not allow the type mangling that is used in Refpolicy.
Instead, a string is passed to the macro and the string rule is
used to create a named string that is used throughout the macro.

Example:
macro foo_template(string $1)
  string foo_template1 {$1 _t};
  string foo_template2 {$1 _tmp_t};
  ...

So the original mangled types $1_t and $1_tmp_t have to be found
and replaced in the macro with foo_template1 and foo_template2.

There was an error in how this was done before. A search was done
for the mangled type in a line and, if found, it is replaced. This
does not work if one mangled type is a subset of another and the
search and replace is done for it first. In the example above,
$1_t is a subset of $1_tmp_t, so if $1_t is searched for first,
the $1_tmp_t instances will be replaced with foo_template1mp_t
which is, obviously, not what is desired.

Instead, check every whitespace separated word in the line and
see if it matches one of the mangled types.

Signed-off-by: James Carter <jwcarter@gmail.com>